Done refactoring all Configuration into its own Class (Marlin3DprinterToolConfiguration)
