+++ Marlin3DprinterTool/Marlin3DprinterTool.sln
@@ -21,6 +21,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpShellServerManager", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Marlin3DprinterToolUserControls", "Marlin3DprinterToolUserControls\Marlin3DprinterToolUserControls.csproj", "{D2627DFC-7450-4F41-8C80-A261BFD52D34}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Marlin3DprinterToolConfiguration", "Marlin3DprinterToolConfiguration\Marlin3DprinterToolConfiguration.csproj", "{A8FE0E27-F9C1-466B-A404-28993E7F75CD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -137,6 +139,18 @@ Global
 		{D2627DFC-7450-4F41-8C80-A261BFD52D34}.Release|x64.Build.0 = Release|Any CPU
 		{D2627DFC-7450-4F41-8C80-A261BFD52D34}.Release|x86.ActiveCfg = Release|Any CPU
 		{D2627DFC-7450-4F41-8C80-A261BFD52D34}.Release|x86.Build.0 = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|x64.Build.0 = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Debug|x86.Build.0 = Debug|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|x64.ActiveCfg = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|x64.Build.0 = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|x86.ActiveCfg = Release|Any CPU
+		{A8FE0E27-F9C1-466B-A404-28993E7F75CD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
