+++ Marlin3DprinterTool/MarlinComunicationHelper/FrmShowCommunication.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Windows.Forms;
 using FastColoredTextBoxNS;
+using Marlin3DprinterToolConfiguration;
 
 namespace MarlinComunicationHelper
 {
