+++ Marlin3DprinterTool/MarlinEditor/FrmMarlinEditorSetup.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Marlin3DprinterToolConfiguration;
 
 namespace MarlinEditor
 {
