+++ Marlin3DprinterTool/MarlinEditor/FrmMigrationCompare.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Text;
 using System.Windows.Forms;
 using FastColoredTextBoxNS;
+using Marlin3DprinterToolConfiguration;
 
 namespace MarlinEditor
 {
