+++ Marlin3DprinterTool/MarlinEditor/MarlinEditor.csproj
@@ -73,7 +73,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ArduinoIDE.cs" />
-    <Compile Include="Configuration.cs" />
     <Compile Include="FrmFirmwareMigrate.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -206,6 +205,12 @@
     <Content Include="MarlinEditorConfiguration.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Marlin3DprinterToolConfiguration\Marlin3DprinterToolConfiguration.csproj">
+      <Project>{a8fe0e27-f9c1-466b-a404-28993e7f75cd}</Project>
+      <Name>Marlin3DprinterToolConfiguration</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
